Shifted is_letter(), is_candidate(), get_letter(), select_alias(), and invert_aliases() from fetch_catalogs.py to catalog_utils.py. Shifted normalize_name() from source_catalog.py to catalog_utils.py. Implemeted get_trexolists_targets() to do extract target names from a trexolists csv file.
